Refactor Cuckoo class to act like a Set (as opposed to a map)
I like the performance of this Cuckoo filter implementation (when compared with
other implementations around, for my particular use cases (where I do not know
even the approximate number of keys I will find, but I know they can be large)).

However, I only need to check for the presence of keys, and never for values.

So, I am dropping the ability of those filters to store values (to decrease RAM
usage and (only a little bit), CPU usage)